Various refactor
