+++ go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	github.com/PuerkitoBio/goquery v1.8.0
 	github.com/aws/aws-lambda-go v1.28.0
-	github.com/aws/aws-sdk-go v1.42.43
+	github.com/aws/aws-sdk-go v1.42.47
 	github.com/awslabs/aws-lambda-go-api-proxy v0.12.0
 	github.com/joho/godotenv v1.4.0
 	github.com/line/line-bot-sdk-go v7.8.0+incompatible
